refactor: split horizon planning in two
